Disabled postprocessor that realigns equal signs of multiple variable assignments.
I once considered this good practice, but reality and experience shows that artificially/manually aligning variable assignments this way does not work out in practice.

Once a new variable is added, or the name of an existing is changed, all of the other lines suddenly need to be changed.  Thus, the versioning system will record changes for code that is otherwise unaffected.  And other patches/changes needlessly run into conflicts due to larger variable assignment block changes.